fix: rename pageOfOperation to operation-form
