+++ src/lib/components/operation-form.svelte
